no message
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@1128 3db68df2-c116-0410-a063-a993310a9797

+++ source/gui/CGUI.cpp
@@ -927,7 +927,7 @@ void CGUI::DrawText(const SGUIText &Text, const CColor &DefaultColor,
 		glPushMatrix();
 
 		// TODO Gee: (2004-09-04) Why are font corrupted if inputted float value?
-		glTranslatef((int)(pos.x+it->m_Pos.x), (int)(pos.y+it->m_Pos.y), z);
+		glTranslatef((GLfloat)int(pos.x+it->m_Pos.x), (GLfloat)int(pos.y+it->m_Pos.y), z);
 		glColor4f(color.r, color.g, color.b, color.a);
 		glwprintf(it->m_String);
